refactor: useDropdown을 useOutsideClick으로 변경해서 다른 ui요소도 외부 클릭 기능을 사용할 수 있도록 변경
